refactor!: remove nested account params for non-alchemy clients
