+++ site/pages/migration-guide.mdx
@@ -38,6 +38,11 @@ This method on the `AccountSigner` has been renamed.
 
 If you were creating LightAccounts for users without explicitly specifying the version (ie. in `createLightAccount` or `useAccount` or `useSmartAccountClient`), you should manually specify the previous default of `v1.1.0`.
 
+### Accounts: `create*Client` methods now have the account parameters un-nested
+
+Previously, the `create*Client` (ie. `createLightAccountClient`) methods used to have an `account` parameter which took in the params for the underlying account. This is resulted in different APIs for both the corresponding `create*AlchemyClient` (ie. `createLightAccountAlchemyClient`) which had these as top-level parameters.
+If you were using `createLightAccountClient` or similar methods, you should now pass the account parameters as top-level parameters.
+
 ## Migrating to version 3.x.x
 
 This version update brings a lot of breaking changes. As we began to support Modular Accounts and their interfaces, we realized that the previous version of the SDK was not as flexible as it could have been to handle the modularity of the new account interfaces.
